feat(scripts): load scripts on head
There's a limitation in the orgmode export where you can only insert
snippets in the head. This was causing some trouble since we have to
query for elements on the DOM in order to support some theme features
such as "back to top".

This commit allows the script to work properly even when injected within
the header element.

Closes #9